Separate Rohde&Schwarz/Hameg driver code from model descriptions.
The number of models/series currently supported by the new
Rohde&Schwarz/Hameg driver ("rohde-schwarz-hameg") has grown
considerably !

Move all model descriptions from the C code file to a new separate
header file.
